Create AuxKernelBase to allow for VectorAuxKernel and AuxKernel classes
(refs #12312)
